requirements.txt
# Web3 & Ethereum
web3>=6.0.0
eth-abi>=4.0.0
eth-utils>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0

# Async Support
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
rich>=13.0.0
httpx>=0.24.0

# Development
ipython>=8.0.0
jupyter>=1.0.0
